@@ -7,7 +7,6 @@
 package apt
 
 import (
-	"context"
 	"testing"
 
 	"github.com/hashicorp/terraform-plugin-go/tftypes"
@@ -103,7 +102,7 @@ func TestStandardRepoHandleValueFromTerraform(t *testing.T) {
 			name, func(t *testing.T) {
 				t.Parallel()
 
-				ctx := context.TODO()
+				ctx := t.Context()
 				val, err := StandardRepoHandleType{}.ValueFromTerraform(ctx, test.val)
 
 				if err == nil && test.expectError {
